+++ server/src/engine/error.rs
@@ -253,5 +253,6 @@ enumerate_err! {
         */
         RawJournalRuntimeDirty = "journal-in-dirty-state",
         RuntimeRestoreValidationFailure = "restore-data-validation-failed",
+        RuntimeEngineLoadError = "runtime-storage-engine-load-failure",
     }
 }